# modules/services/src/main/resources/openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "LEOS SERVICE API",
    "description" : "API documentation for LEOS core services",
    "version" : "5.5.0-SNAPSHOT"
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/secured/document-ref/{packageId}" : {
      "get" : {
        "tags" : [ "LEOS API" ],
        "summary" : "Find document reference by package ID and category",
        "description" : "Retrieves document reference using package ID and category",
        "operationId" : "findDocumentRefByPackageIdAndCategory",
        "parameters" : [ {
          "name" : "packageId",
          "in" : "path",
          "description" : "Package ID",
          "required" : true,
          "schema" : {
            "type" : "string",
            "exampleSetFlag" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Document reference retrieved successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Object",
                  "exampleSetFlag" : false
                },
                "exampleSetFlag" : false
              }
            }
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "HTTP",
        "description" : "JWT Bearer token authentication",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "basicAuth" : {
        "type" : "HTTP",
        "description" : "Basic HTTP authentication",
        "scheme" : "basic"
      }
    }
  }
}